refactor: extract cardCamel and getChatString into shared chat-utils module
